created my submission folder for text diffusion
