+++ requirements.txt
@@ -8,4 +8,3 @@ typing-extensions==4.15.0
 datasets
 tiktoken
 sentencepiece
-zstandard
